Totally done CategoryController

+++ src/main/java/com/itgroup/controllers/AuthorizationController.java
@@ -22,12 +22,12 @@ public class AuthorizationController {
     }
 
     @PatchMapping("/{id}")
-    public String update(@PathVariable("id") int id) {
+    public String update(@PathVariable("id") Long id) {
         return null;
     }
 
     @GetMapping("/{id}")
-    public String show(@PathVariable("id") int id) {
+    public String show(@PathVariable("id") Long id) {
         return null;
     }
 }
